pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.120.0",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.1",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.2.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
